+++ packages/expo-updates/android/src/main/java/expo/modules/updates/errorrecovery/ErrorRecovery.kt
@@ -5,7 +5,7 @@ import android.os.Handler
 import android.os.HandlerThread
 import android.util.Log
 import com.facebook.react.ReactInstanceManager
-import com.facebook.react.bridge.DefaultNativeModuleCallExceptionHandler
+import com.facebook.react.bridge.DefaultJSExceptionHandler
 import com.facebook.react.bridge.ReactMarker
 import com.facebook.react.bridge.ReactMarkerConstants
 import com.facebook.react.devsupport.DisabledDevSupportManager
@@ -22,7 +22,7 @@ class ErrorRecovery(
   internal val logger = UpdatesLogger(context)
 
   private var weakReactInstanceManager: WeakReference<ReactInstanceManager>? = null
-  private var previousExceptionHandler: DefaultNativeModuleCallExceptionHandler? = null
+  private var previousExceptionHandler: DefaultJSExceptionHandler? = null
 
   fun initialize(delegate: ErrorRecoveryDelegate) {
     if (!::handler.isInitialized) {
@@ -72,17 +72,17 @@ class ErrorRecovery(
     }
 
     val devSupportManager = reactInstanceManager.devSupportManager as DisabledDevSupportManager
-    val defaultNativeModuleCallExceptionHandler = object : DefaultNativeModuleCallExceptionHandler() {
+    val defaultJSExceptionHandler = object : DefaultJSExceptionHandler() {
       override fun handleException(e: Exception?) {
         this@ErrorRecovery.handleException(e!!)
       }
     }
     val devSupportManagerClass = devSupportManager.javaClass
-    previousExceptionHandler = devSupportManagerClass.getDeclaredField("mDefaultNativeModuleCallExceptionHandler").let { field ->
+    previousExceptionHandler = devSupportManagerClass.getDeclaredField("mDefaultJSExceptionHandler").let { field ->
       field.isAccessible = true
       val previousValue = field[devSupportManager]
-      field[devSupportManager] = defaultNativeModuleCallExceptionHandler
-      return@let previousValue as DefaultNativeModuleCallExceptionHandler
+      field[devSupportManager] = defaultJSExceptionHandler
+      return@let previousValue as DefaultJSExceptionHandler
     }
     weakReactInstanceManager = WeakReference(reactInstanceManager)
   }
@@ -99,7 +99,7 @@ class ErrorRecovery(
 
       val devSupportManager = reactInstanceManager.devSupportManager as DisabledDevSupportManager
       val devSupportManagerClass = devSupportManager.javaClass
-      devSupportManagerClass.getDeclaredField("mDefaultNativeModuleCallExceptionHandler").let { field ->
+      devSupportManagerClass.getDeclaredField("mDefaultJSExceptionHandler").let { field ->
         field.isAccessible = true
         field[devSupportManager] = previousExceptionHandler
       }