+++ packages/jest-expo/CHANGELOG.md
@@ -12,6 +12,10 @@
 
 ### 💡 Others
 
+### ⚠️ Notices
+
+- Added support for React Native 0.70.x. ([#19261](https://github.com/expo/expo/pull/19261) by [@kudo](https://github.com/kudo))
+
 ## 46.0.1 — 2022-07-11
 
 _This version does not introduce any user-facing changes._